requirements.txt
openai>=1.0
livekit>=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector,google]~=1.0